backend/pyproject.toml
[project]
name = "assessment"
version = "0.1.0"
description = "Multi-agent AI compliance system for fraud alert processing"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.3",
    "fastapi[standard]>=0.135.2",
    "pgvector>=0.4.2",
    "pydantic>=2.12.5",
    "langgraph>=0.5.0",
    "langchain-google-vertexai>=2.0.0",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    "google-cloud-bigquery>=3.27.0",
    "google-cloud-storage>=2.19.0",
    "google-cloud-aiplatform>=1.75.0",
    "langfuse>=2.58.0",
    "psycopg[binary]>=3.2.0",
    "sentence-transformers>=3.4.0",
    "pypdf>=5.3.0",
    "tenacity>=9.1.0",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "ruff>=0.15.7",
    "pytest>=8.3.0",
    "pytest-django>=4.10.0",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.0",
    "httpx>=0.28.0",
    "factory-boy>=3.3.0",
    "freezegun>=1.5.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.local"
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
addopts = "--cov=compliance_agent --cov-report=term-missing"
markers = [
    "integration: marks tests as integration tests (require external services)",
]

[tool.ruff]
line-length = 88
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "S", "B", "C4", "RET", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S108", "S311", "S608"]
"compliance_agent/tools/bigquery_tools.py" = ["S608", "S311"]
"compliance_agent/api/routers/alerts.py" = ["B008"]
